Iterate linear regression across all combinations of two matrices

Iterate linear regression across all combinations of columns between two matrices, for example cell types (xmat) and phenotypes (ymat).

iterate_lm(
  xmat,
  ymat,
  test_method = c("glm", "glm_univariate", "anova", "lm.ridge", "rlm"),
  multivariate = FALSE,
  scale_fn = NULL,
  correction_method = "fdr",
  qvalue_thresh = 0.05,
  quantize = list(x = NULL, y = NULL),
  progressbar = TRUE,
  workers = NULL,
  verbose = TRUE,
  ...
)

Arguments

xmat

gene x trait matrix.

ymat

gene x celltype matrix.

test_method

Association testing method to use.

multivariate

If TRUE, runs tests with each column in xmat as a multivariate predictor in a single model (one model per column in yvar). IfFALSE, runs tests with each column in xmat as a univariate predictor in separate models (one model per column in yvar).

scale_fn

A function to scale the x and y variables with before running the regression. This helps to make the coefficients comparable across different models. Set to NULL to skip the scaling step.

correction_method

Multiple-testing correction method to be passed to p.adjust.

qvalue_thresh

q-value threshold to use when report significant results summary.

quantize

A named list where the values of "x" and "y" indicate the number of quantiles to bin the respective xmat and ymat datasets into.

progressbar

logical(1) Enable progress bar (based on plyr:::progress_text). Enabling the progress bar changes the default value of tasks to .Machine$integer.max, so that progress is reported for each element of X.

workers

Number (>1) or proportion (<1) of worker cores to use.

verbose

Print messages.

...

Additional parameters passed to the statistical test function.

Examples

set.seed(42)

### x matrix (e.g. cell type expression specificity signatures)
xmat <- simulate_matrix(nrow = 200, ncol = 10, col_prefix="celltype_")

### y matrix (e.g. gene-phenotype association signatures)
ymat <- simulate_matrix(nrow = 100, ncol = 50, col_prefix="phenotype_") 

res_lm <- iterate_lm(xmat = xmat,
                     ymat = ymat,
                     test_method = "glm_univariate",
                     workers = 1)
#> 1 core(s) assigned as workers (3 reserved).
#> 100 intersecting features between all matrices.
#> Running 500 tests: 10 xmat columns x 50 ymat columns.
#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   2%  |                                                                              |===                                                                   |   4%  |                                                                              |====                                                                  |   6%  |                                                                              |======                                                                |   8%  |                                                                              |=======                                                               |  10%  |                                                                              |========                                                              |  12%  |                                                                              |==========                                                            |  14%  |                                                                              |===========                                                           |  16%  |                                                                              |=============                                                         |  18%  |                                                                              |==============                                                        |  20%  |                                                                              |===============                                                       |  22%  |                                                                              |=================                                                     |  24%  |                                                                              |==================                                                    |  26%  |                                                                              |====================                                                  |  28%  |                                                                              |=====================                                                 |  30%  |                                                                              |======================                                                |  32%  |                                                                              |========================                                              |  34%  |                                                                              |=========================                                             |  36%  |                                                                              |===========================                                           |  38%  |                                                                              |============================                                          |  40%  |                                                                              |=============================                                         |  42%  |                                                                              |===============================                                       |  44%  |                                                                              |================================                                      |  46%  |                                                                              |==================================                                    |  48%  |                                                                              |===================================                                   |  50%  |                                                                              |====================================                                  |  52%  |                                                                              |======================================                                |  54%  |                                                                              |=======================================                               |  56%  |                                                                              |=========================================                             |  58%  |                                                                              |==========================================                            |  60%  |                                                                              |===========================================                           |  62%  |                                                                              |=============================================                         |  64%  |                                                                              |==============================================                        |  66%  |                                                                              |================================================                      |  68%  |                                                                              |=================================================                     |  70%  |                                                                              |==================================================                    |  72%  |                                                                              |====================================================                  |  74%  |                                                                              |=====================================================                 |  76%  |                                                                              |=======================================================               |  78%  |                                                                              |========================================================              |  80%  |                                                                              |=========================================================             |  82%  |                                                                              |===========================================================           |  84%  |                                                                              |============================================================          |  86%  |                                                                              |==============================================================        |  88%  |                                                                              |===============================================================       |  90%  |                                                                              |================================================================      |  92%  |                                                                              |==================================================================    |  94%  |                                                                              |===================================================================   |  96%  |                                                                              |===================================================================== |  98%  |                                                                              |======================================================================| 100%
#> 
#> 0 significant results @ fdr < 0.05
#> Time difference of 2.382318 secs